refactor: views 클래스 시작점 생성
- 가독성을 위해 view에 해당하는 클래스들을 index.js를 만들어주어 관리해주도록 변경

- 관련 피드백: https://github.com/woowacourse/javascript-lotto/pull/136#discussion_r820087835
